fix EID chipid issue

+++ Helper.ino
@@ -10,7 +10,7 @@
 
 #include "rom/rtc.h"
 
-#define WIFI_getChipId() (uint32_t)ESP.getEfuseMac()
+#define WIFI_getChipId() (uint64_t)ESP.getEfuseMac()
 
 const PROGMEM char *resetReasons[]  { "Unknown", "Vbat power on reset", "2-unknown","Software reset digital core", "Legacy watch dog reset digital core", 
 "Deep Sleep reset digital core", "Reset by SLC module, reset digital core","Timer Group0 Watch dog reset digital core","Timer Group1 Watch dog reset digital core",
